A looooong day for frontend begins
i believe the updates work now
drag to reorder and all that
but now im in the create ranking page which was worked on before i had a backend lmao so it was missing some stuff

+++ simple-rank-backend/Models/RankItem.cs
@@ -16,11 +16,11 @@ namespace simple_rank_backend.Models
             ItemId = string.Empty;
         }
 
-        public RankItem(string name, string description, uint rank, string itemId)
+        public RankItem(string name, string description, int rank, string itemId)
         {
             Name = name;
             Description = description;
-            Rank = rank;
+            Rank = (uint)rank;
             ItemId = itemId;
         }
